minor refactoring + file duplicate check logic implemented
